requirements.txt
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.30.0
diffusers>=0.20.0
ultralytics>=8.0.0
opencv-python>=4.8.0
Pillow>=10.0.0
numpy>=1.24.0
accelerate>=0.20.0
safetensors>=0.3.0
segment-anything-py>=1.0
requests>=2.31.0
tqdm>=4.65.0
matplotlib>=3.7.0
fastapi>=0.104.0
uvicorn>=0.24.0
python-multipart>=0.0.6
pydantic>=2.4.0
# Additional dependencies for robust AI models
huggingface-hub>=0.16.0
datasets>=2.14.0
tokenizers>=0.13.0
scipy>=1.11.0
scikit-learn>=1.3.0
scikit-image>=0.21.0
rembg>=2.0.50
